bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["IamIknow <stepanov.serge97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.0a6"
python-dotenv = "^0.21.0"
aiokafka = "^0.8.0"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
